Removed get_next_line directory and associated unused functions, moved ft_strrejoin() and copy_and_cat() to here_doc_utils.c, functions declared in minishell.h
